Remove destruct BaseAmqp
Close connection is already managed in `destruct` method of `AMQPConnection` class

If we keep this destruct, it will be called first and will close connection.
Then `AMQPConnection` destruct method will be called, and if `close_on_destruct` property is true, will try to close again... and we will get an exception.